Move RocksDbConfig back to server, pass byte sizes to DatabaseManager
Revert the previous move of `RocksDbConfig` into `common/options` —
it belongs alongside the rest of the server config tree in
`server/parameters/config.rs`, not as a primitive in the options
crate.

`DatabaseManager::new` now takes the two `ByteSize` values directly
(`rocksdb_cache_size`, `rocksdb_write_buffers_limit`) instead of a
`&RocksDbConfig` reference. This keeps the database crate ignorant
of the server's config types and makes the call site at
`server/state/mod.rs` explicit about what's being threaded down.

Test/bench call sites that previously constructed `&RocksDbConfig::
default()` now pass `ByteSize::mb(64), ByteSize::mb(64)` directly,
matching the original 64 MiB-each test budgets.

Also fix a small follow-up: `database/tests/database.rs:15` was the
last `64 * 1024 * 1024` magic number that wasn't migrated to `64 *
MB`. Adding `//resource` to its `test_database` Bazel target lets it
import `resource::constants::common::MB` like the other migrated
sites.
