+++ database/tests/statistics_synchronization.rs
@@ -20,7 +20,7 @@ use database::{
 };
 use diagnostics::diagnostics_manager::DiagnosticsManager;
 use executor::ExecutionInterrupt;
-use options::{QueryOptions, RocksDbConfig, TransactionOptions};
+use options::{ByteSize, QueryOptions, TransactionOptions};
 use query::given_rows::GivenRowsSimple;
 use storage::durability_client::WALClient;
 use test_utils::{create_tmp_storage_dir, init_logging};
@@ -48,9 +48,13 @@ fn statistics_synchronization_under_concurrent_load() {
     let total_has = 2 * total_persons;
 
     {
-        let dbm =
-            DatabaseManager::new(&tmp_dir, Arc::new(DiagnosticsManager::new_disabled()), &RocksDbConfig::default())
-                .unwrap();
+        let dbm = DatabaseManager::new(
+            &tmp_dir,
+            Arc::new(DiagnosticsManager::new_disabled()),
+            ByteSize::mb(64),
+            ByteSize::mb(64),
+        )
+        .unwrap();
         dbm.put_database(DB_NAME).unwrap();
         let database = dbm.database(DB_NAME).unwrap();
 
@@ -79,8 +83,13 @@ fn statistics_synchronization_under_concurrent_load() {
 
     // dbm and database dropped here; IntervalRunner threads shut down synchronously on drop.
 
-    let dbm = DatabaseManager::new(&tmp_dir, Arc::new(DiagnosticsManager::new_disabled()), &RocksDbConfig::default())
-        .unwrap();
+    let dbm = DatabaseManager::new(
+        &tmp_dir,
+        Arc::new(DiagnosticsManager::new_disabled()),
+        ByteSize::mb(64),
+        ByteSize::mb(64),
+    )
+    .unwrap();
     let database = dbm.database(DB_NAME).unwrap();
     let metrics = database.get_metrics();
