+++ database/tests/transaction.rs
@@ -14,7 +14,7 @@ use database::{
     transaction::{CommitIntent, TransactionRead, TransactionSchema, TransactionWrite},
 };
 use diagnostics::diagnostics_manager::DiagnosticsManager;
-use options::{RocksDbConfig, TransactionOptions};
+use options::{ByteSize, TransactionOptions};
 use storage::durability_client::WALClient;
 use test_utils::{TempDir, create_tmp_storage_dir, init_logging};
 use tokio::{
@@ -39,9 +39,13 @@ macro_rules! assert_transaction_timeout {
 }
 
 fn create_database(databases_path: &TempDir) -> Arc<Database<WALClient>> {
-    let database_manager =
-        DatabaseManager::new(databases_path, Arc::new(DiagnosticsManager::new_disabled()), &RocksDbConfig::default())
-            .expect("Expected database manager");
+    let database_manager = DatabaseManager::new(
+        databases_path,
+        Arc::new(DiagnosticsManager::new_disabled()),
+        ByteSize::mb(64),
+        ByteSize::mb(64),
+    )
+    .expect("Expected database manager");
     database_manager.put_database(DB_NAME).expect("Expected database creation");
     database_manager.database(DB_NAME).expect("Expected database retrieval")
 }
