+++ tests/benchmarks/iam/bench_iam.rs
@@ -13,7 +13,7 @@ use database::{
 };
 use diagnostics::diagnostics_manager::DiagnosticsManager;
 use executor::{ExecutionInterrupt, batch::Batch, pipeline::stage::StageIterator};
-use options::{RocksDbConfig, TransactionOptions};
+use options::{ByteSize, TransactionOptions};
 use query::given_rows::GivenRowsSimple;
 use storage::durability_client::WALClient;
 use test_utils::create_tmp_storage_dir;
@@ -113,9 +113,13 @@ fn load_data_tql(database: Arc<Database<WALClient>>, data_tql: &Path) {
 fn setup() -> Arc<Database<WALClient>> {
     let tmp_dir = create_tmp_storage_dir();
     {
-        let dbm =
-            DatabaseManager::new(&tmp_dir, Arc::new(DiagnosticsManager::new_disabled()), &RocksDbConfig::default())
-                .unwrap();
+        let dbm = DatabaseManager::new(
+            &tmp_dir,
+            Arc::new(DiagnosticsManager::new_disabled()),
+            ByteSize::mb(64),
+            ByteSize::mb(64),
+        )
+        .unwrap();
         dbm.put_database(DB_NAME).unwrap();
         let database = dbm.database(DB_NAME).unwrap();
         let schema_path = Path::new(RESOURCE_PATH).join(Path::new(SCHEMA_FILENAME));
@@ -126,8 +130,13 @@ fn setup() -> Arc<Database<WALClient>> {
         load_schema_tql(database.clone(), &functions_path);
         load_data_tql(database.clone(), &data_path);
     }
-    let dbm = DatabaseManager::new(&tmp_dir, Arc::new(DiagnosticsManager::new_disabled()), &RocksDbConfig::default())
-        .unwrap();
+    let dbm = DatabaseManager::new(
+        &tmp_dir,
+        Arc::new(DiagnosticsManager::new_disabled()),
+        ByteSize::mb(64),
+        ByteSize::mb(64),
+    )
+    .unwrap();
     dbm.put_database(DB_NAME).unwrap();
 
     dbm.database(DB_NAME).unwrap()
